MajorizationFigures

using MajorizationFigures, Polyhedra
using MajorizationFigures: E
using MajorizationExtrema
using Combinatorics

q = [21 // 100, 24 // 100, 55 // 100]
ϵ = 1 // 10
Bq = TV_Ball(q, ϵ)
Es = E.(permutations(1:3))
figure(Es...)

polys = Polyhedron[]
push!(polys, E((1, 2, 3)))
append!(polys, sort(Bq; rev = true))
append!(polys, [Bq ∩ s for s in Es])
figure(q, polys...)

figure(q, majmin(q, ϵ), polys...)

figure(q, sort(majmin(q, ϵ); rev = true), polys...)

figure(q)

Reference

Base.sort

MajorizationFigures.E

Base.sort — Method

Base.sort(p::Polyhedron; kwargs...) -> Vector{Polyhedron}

Returns a collection of polyhedra whose union gives the set equal to sorting each element in p elementwise.

MajorizationFigures.E — Method

E(perm) -> Polyedron

Returns the polyhedron consisting of the subset of the probability simplex that is sorted according to the permutation perm, in the sense that for q ∈ E(perm), issorted(q[perm]; rev=true) == true.